tests: Move live_migration to integration_live_migration.rs
Extract mod live_migration (including live_migration_parallel
and live_migration_sequential submodules) from integration.rs
into a standalone test binary integration_live_migration.rs.
This reduces the size of integration.rs and follows the
pattern established by integration_cvm.rs.
