+++ cloud-hypervisor/tests/common/mod.rs
@@ -2,5 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod live_migration;
 pub(crate) mod tests_wrappers;
 pub(crate) mod utils;
